fix user's avatar part; small improvements of code;
