+++ src/main/java/ru/skypro/homework/controller/ImageController.java
@@ -1,15 +1,10 @@
 package ru.skypro.homework.controller;
 
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.service.ImageService;
 
 @Slf4j
@@ -20,25 +15,6 @@ import ru.skypro.homework.service.ImageService;
 public class ImageController {
 
     private  final ImageService imageService;
-
-    @Operation(summary = "updateAdsImage",
-            responses = {
-                    @ApiResponse(
-                            responseCode = "200",
-                            content = @Content(
-                                    mediaType = MediaType.APPLICATION_OCTET_STREAM_VALUE,
-                                    schema = @Schema(implementation = byte[].class)
-                            )
-                    ),
-                    @ApiResponse(responseCode = "404", content = @Content)
-            })
-    @PatchMapping (value = "{id}", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<byte[]> updateAdsImage(@PathVariable Long id, @RequestParam MultipartFile image){
-        log.info("Was invoked updateAdsImage method from {}", ImageController.class.getSimpleName());
-
-        byte[] imageBytes = imageService.updateAdsImage(id, image);
-        return ResponseEntity.ok(imageBytes);
-    }
 
     @GetMapping(value = "{id}", produces = {MediaType.IMAGE_PNG_VALUE})
     public ResponseEntity<byte[]> getAdsImage(@PathVariable Long id) {
