+++ src/main/java/ru/skypro/homework/mapper/UserMapper.java
@@ -1,11 +1,9 @@
 package ru.skypro.homework.mapper;
 
-import org.mapstruct.InjectionStrategy;
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import org.mapstruct.MappingConstants;
+import org.mapstruct.*;
 import ru.skypro.homework.dto.ResponseWrapperUserDto;
 import ru.skypro.homework.dto.UserDto;
+import ru.skypro.homework.entity.Avatar;
 import ru.skypro.homework.entity.User;
 
 import java.util.List;
@@ -15,6 +13,7 @@ import java.util.List;
 public interface UserMapper {
 
     @Mapping(source = "username", target = "email")
+    @Mapping(source = "avatar", target = "image", qualifiedByName = "getUserAvatarLink")
     UserDto userToUserDto(User user);
 
     @Mapping(source = "email", target = "username")
@@ -23,5 +22,10 @@ public interface UserMapper {
     @Mapping(source = "size", target = "count")
     @Mapping(source = "userList", target = "results")
     ResponseWrapperUserDto userListToResponseWrapperUserDto(Integer size,List<User> userList);
+
+    @Named("getUserAvatarLink")
+    default String getUserAvatarLink(Avatar avatar) {
+        return (avatar == null) ? null : "/users/avatar/" + avatar.getId();
+    }
 
 }
